py_llm/pyproject.toml
[project]
name = "py-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub>=1.2.4",
    "llama-cpp-python>=0.3.16",
]
